Fix issue with logic around type of object returned by smart_open
This fixes an issue where the return type of smart_open can vary in
different environments. There was a check in place that was made under
the assumption the return type would be consistent with what I was
seeing in my development environment at time of coding; however in
another instance I found this open returned an unexpected type for an
open file.

This change avoids having type checking interefere with runtime logic
by assigning the open for file to `Any`, which skips further type checking.