Include pgstac migrations in pypgstac package (#108)
* Include pgstac migrations in pypgstac package

* Add tests for specific invalid asset regression

We've had both db and api regressions where assets would be included in
the hydrated item without an href causing validation errors. Adding
specific tests for this case to make sure the general do-no-merge key
solution address this problem, and the api-fix can be removed from
stac-fastapi backend.

* Clarify comments and variable names

* On the recursive portion of the dehydrate function, base_item and
full_item are not actual items, but sub-keys from root. Move to a
private inner function with clearer names, similar to hydrate.
* Update comments to be more precise with respect to leaving items on
the item or not.