Ensure special keys are not in content when loaded (#112)
* Ensure special keys not in content when loaded

The loader pop'd keys like collection off the item, then dehydrated the
item to be used as the 'content'. With collection removed prior to
dehydration, it was flagged with the "do-not-merge" marker because the
key is on the base_item. Instead, ensure that id, collection, and
geometry are not in content as they are stored on the table row and
shouldn't participate in hydration.

Additionally, bbox was previously a derived value at search runtime but
was recently changed to a returned value if it existed on the item.
However, during loading, the bbox was dropped so would never exist on
the persisted item.

* Simplify geometry check