Fix collection check in queryable trigger_func (#179)
* Fix collection existence check in queryables

The previous check for the existence of a collection could still return values
that weren't in the new collection_ids array. This issue was only
present when there were more collections than just the collection being
tested. I added an additional test collection, added a test to verify
the error and then updated the sql join to only return results if there
is a missing collection.

* Update queryable docs to account for text[]

Use a more specific query to correctly insert queryables from
`missing_queryables`

* migration

* Changelog

---------

Co-authored-by: David Bitner <bitner@dbspatial.com>