Remove default eo:cloud_cover queryable
This default queryable was added in the early days of pgstac so that the
field would be indexed by default if it was present in any items.
However, now that the queryables table is used to generate content for
/queryables stac endpoint, it is included in the queryable content for
all collections (since there is no collection_id).

Removing this opmtimization fixes incorrect (and broken, since the $ref
url also has a typo) default behavior.